Revert "Revert "Close old connections before attempting to store task state""
